LOB_LOCATOR_STATE to elo.h
