+++ src/communication/network_interface_cl.h
@@ -34,6 +34,7 @@
 #include <stdio.h>
 
 #include "dbtype_def.h"
+#include "elo.h"
 #include "replication.h"
 #include "server_interface.h"
 #include "perf_monitor.h"
