+++ src/object/elo.h
@@ -31,6 +31,44 @@
 #include <stdio.h>
 #include "object_representation.h"
 
+/* there can be following transitions in transient lobs
+
+   -------------------------------------------------------------------------
+   | 	       locator  | created               | deleted		   |
+   |--------------------|-----------------------|--------------------------|
+   | in     | transient | LOB_TRANSIENT_CREATED i LOB_UNKNOWN		   |
+   | tran   |-----------|-----------------------|--------------------------|
+   |        | permanent | LOB_PERMANENT_CREATED | LOB_PERMANENT_DELETED    |
+   |--------------------|-----------------------|--------------------------|
+   | out of | transient | LOB_UNKNOWN		| LOB_UNKNOWN		   |
+   | tran   |-----------|-----------------------|--------------------------|
+   |        | permanent | LOB_UNKNOWN 		| LOB_TRANSIENT_DELETED    |
+   -------------------------------------------------------------------------
+
+   s1: create a transient locator and delete it
+       LOB_TRANSIENT_CREATED -> LOB_UNKNOWN
+
+   s2: create a transient locator and bind it to a row in table
+       LOB_TRANSIENT_CREATED -> LOB_PERMANENT_CREATED
+
+   s3: bind a transient locator to a row and delete the locator
+       LOB_PERMANENT_CREATED -> LOB_PERMANENT_DELETED
+
+   s4: delete a locator to be create out of transaction
+       LOB_UNKNOWN -> LOB_TRANSIENT_DELETED
+
+ */
+enum lob_locator_state
+{
+  LOB_UNKNOWN,
+  LOB_TRANSIENT_CREATED,
+  LOB_TRANSIENT_DELETED,
+  LOB_PERMANENT_CREATED,
+  LOB_PERMANENT_DELETED,
+  LOB_NOT_FOUND
+};
+typedef enum lob_locator_state LOB_LOCATOR_STATE;
+
 #ifdef __cplusplus
 extern "C"
 {
